requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
aiosqlite==0.21.0
attrs==25.3.0
discord.py==2.4.0
frozenlist==1.7.0
idna==3.10
multidict==6.6.4
propcache==0.3.2
typing_extensions==4.15.0
yarl==1.20.1
